refactor(out format): use short file names.
