+++ src/query/formats/src/field_encoder/json.rs
