+++ src/query/formats/src/field_encoder/mod.rs
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod field_encoder_csv;
-mod field_encoder_json;
-mod field_encoder_row_based;
-mod field_encoder_tsv;
-mod field_encoder_values;
+mod csv;
 mod helpers;
+mod json;
+mod row_based;
+mod tsv;
+mod values;
 
-pub use field_encoder_csv::FieldEncoderCSV;
-pub use field_encoder_json::FieldEncoderJSON;
-pub use field_encoder_row_based::FieldEncoderRowBased;
-pub use field_encoder_tsv::FieldEncoderTSV;
-pub use field_encoder_values::FieldEncoderValues;
+pub use csv::FieldEncoderCSV;
 pub use helpers::CommonSettings;
+pub use json::FieldEncoderJSON;
+pub use row_based::FieldEncoderRowBased;
+pub use tsv::FieldEncoderTSV;
+pub use values::FieldEncoderValues;
